+++ component/prometheus/scrape/scrape.go
@@ -16,6 +16,7 @@ import (
 	"github.com/grafana/agent/component/discovery"
 	"github.com/grafana/agent/component/prometheus"
 	"github.com/grafana/agent/pkg/build"
+	client_prometheus "github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/config"
 	"github.com/prometheus/prometheus/discovery/targetgroup"
@@ -107,10 +108,11 @@ type Component struct {
 
 	reloadTargets chan struct{}
 
-	mut        sync.RWMutex
-	args       Arguments
-	scraper    *scrape.Manager
-	appendable *prometheus.Fanout
+	mut          sync.RWMutex
+	args         Arguments
+	scraper      *scrape.Manager
+	appendable   *prometheus.Fanout
+	targetsGauge *client_prometheus.GaugeVec
 }
 
 var (
@@ -127,6 +129,7 @@ func New(o component.Options, args Arguments) (*Component, error) {
 		reloadTargets: make(chan struct{}, 1),
 		scraper:       scraper,
 		appendable:    flowAppendable,
+		targetsGauge:  getTargetsGaugeMetric(o.Registerer),
 	}
 
 	// Call to Update() to set the receivers and targets once at the start.
@@ -200,6 +203,7 @@ func (c *Component) Update(args component.Arguments) error {
 	default:
 	}
 
+	c.targetsGauge.WithLabelValues().Set(float64(len(c.args.Targets)))
 	return nil
 }
 
@@ -294,4 +298,21 @@ func convertLabelSet(tg discovery.Target) model.LabelSet {
 		lset[model.LabelName(k)] = model.LabelValue(v)
 	}
 	return lset
+}
+
+func getTargetsGaugeMetric(registerer client_prometheus.Registerer) *client_prometheus.GaugeVec {
+	targetsGauge := client_prometheus.NewGaugeVec(client_prometheus.GaugeOpts{
+		Name: "agent_prometheus_scrape_targets_gauge",
+		Help: "A gauge of all scrape targets this component is configured to scrape",
+	}, []string{})
+	err := registerer.Register(targetsGauge)
+	if err != nil {
+		if existing, ok := err.(client_prometheus.AlreadyRegisteredError); ok {
+			targetsGauge = existing.ExistingCollector.(*client_prometheus.GaugeVec)
+		} else {
+			// Same behavior as MustRegister if the error is not for AlreadyRegistered
+			panic(err)
+		}
+	}
+	return targetsGauge
 }